# Improved image gradient GetValue(s) performance by using cached image data